# openapi.yaml
openapi: 3.1.0
info:
  title: Hookflow API
  description: |
    Enterprise-grade distributed webhook delivery system with at-least-once guarantees.

    ## Features
    - **Event Ingestion**: Send events via REST API with automatic fan-out to subscribed endpoints
    - **Reliable Delivery**: Automatic retries with exponential backoff (1m → 24h, 7 attempts)
    - **HMAC Signatures**: Secure webhook verification with SHA-256 signatures
    - **Multi-tenant**: Organization-based isolation with JWT authentication
    - **Rate Limiting**: Distributed rate limiting via Redis

    ## Authentication
    - **JWT Bearer Token**: For all operations (UI / user context)
    - **API Key**: For all project-scoped operations (`X-API-Key` header, SDK usage)
  contact:
    name: Hookflow
    url: https://github.com/vadymkykalo/webhook-platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: "{baseUrl}"
  description: Your Hookflow installation — the same origin the dashboard runs on
  variables:
    baseUrl:
      description: "Scheme and host of your installation, without a trailing slash"
      default: http://localhost:8080
security:
- bearerAuth: []
...
paths:
...
  /api/v1/orgs/{orgId}/members:
    get:
      tags:
      - Organizations
      summary: List members
      description: Returns all members of the organization
      operationId: getMembers
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberResponse"
      security:
      - bearerAuth: []
      - apiKey: []
    post:
      tags:
      - Organizations
      summary: Add member
      description: Invites a user to the organization
      operationId: addMember
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
        required: true
      responses:
        "201":
          description: Member added
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MemberResponse"
      security:
      - bearerAuth: []
      - apiKey: []
  /api/v1/orgs/{orgId}/members/{userId}/suspend:
    post:
      tags:
      - Organizations
      summary: Suspend member
      description: "Suspends a member: the membership and its role are kept, the member\
        \ is refused access, and their current sessions end immediately"
      operationId: suspendMember
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Member suspended
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MemberResponse"
      security:
      - bearerAuth: []
      - apiKey: []
  /api/v1/orgs/{orgId}/members/{userId}/reinstate:
    post:
      tags:
      - Organizations
      summary: Reinstate member
      description: "Lifts a suspension, restoring the member's access in the role\
        \ they kept"
      operationId: reinstateMember
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Member reinstated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MemberResponse"
      security:
      - bearerAuth: []
      - apiKey: []
  /api/v1/orgs/{orgId}/members/accept-invite:
    post:
      tags:
      - Organizations
      summary: Accept invite
      description: Accepts an organization membership invite using the invite token
      operationId: acceptInvite
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: token
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Invite accepted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MemberResponse"
      security:
      - bearerAuth: []
      - apiKey: []
...
  /api/v1/orgs/{orgId}/members/{userId}:
    delete:
      tags:
      - Organizations
      summary: Remove member
      description: Removes a member from the organization
      operationId: removeMember
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Member removed
      security:
      - bearerAuth: []
      - apiKey: []
    patch:
      tags:
      - Organizations
      summary: Change member role
      description: "Updates a member's role (OWNER, ADMIN, MEMBER, VIEWER)"
      operationId: changeMemberRole
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeMemberRoleRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MemberResponse"
      security:
      - bearerAuth: []
      - apiKey: []
...
components:
  schemas:
...
    AddMemberRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        role:
          type: string
          enum:
          - OWNER
          - DEVELOPER
          - VIEWER
          - API_KEY
      required:
      - email
      - role
    MemberResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
          enum:
          - OWNER
          - DEVELOPER
          - VIEWER
          - API_KEY
        status:
          type: string
          enum:
          - INVITED
          - ACTIVE
          - DISABLED
        createdAt:
          type: string
          format: date-time
...
    ChangeMemberRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum:
          - OWNER
          - DEVELOPER
          - VIEWER
          - API_KEY
      required:
      - role
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT token obtained from /api/v1/auth/login
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      description: Project API key for event ingestion
      name: X-API-Key
      in: header
    platformAdminToken:
      type: apiKey
      description: "Cluster-operator credential (PLATFORM_ADMIN_TOKEN env var), independent\
        \ of tenant org membership — required for cross-tenant admin endpoints"
      name: X-Platform-Admin-Token
      in: header